+++ tests/Tests/Fresh.hs
@@ -1,11 +1,36 @@
-module Tests.Fresh where
+{-# LANGUAGE NoImplicitPrelude #-}
+module Tests.Fresh (tests)
+  where
 
-import Control.Monad
-import Control.Monad.Freer
-import Control.Monad.Freer.Fresh
+import Prelude ((-))
+
+import Control.Monad (replicateM)
+import Data.Eq ((==))
+import Data.Function (($), (.))
+import Data.Functor ((<$>))
+import Data.Int (Int)
+import Data.List (last)
+import Data.Ord ((>))
+import Data.Tuple (fst)
+
+import Test.Tasty (TestTree, testGroup)
+import Test.Tasty.HUnit ((@?=), testCase)
+import Test.Tasty.QuickCheck ((==>), testProperty)
+
+import Control.Monad.Freer (Eff, run)
+import Control.Monad.Freer.Fresh (fresh, runFresh)
+
+
+tests :: TestTree
+tests = testGroup "Fresh tests"
+    [ testCase "Start at 0, refresh twice, yields 1"
+        $ testFresh 10 @?= 9
+    , testProperty "Freshening n times yields (n-1)"
+        $ \n -> n > 0 ==> testFresh n == (n-1)
+    ]
 
 makeFresh :: Int -> Eff r Int
-makeFresh n = runFresh' (fmap last (replicateM n fresh)) 0
+makeFresh n = fst <$> runFresh (last <$> replicateM n fresh) 0
 
 testFresh :: Int -> Int
 testFresh = run . makeFresh